Refactor: Moved commands/conanconf to cruizlib
